Added status codes in CGIs, first response string is Status now

+++ nginx_example/cgi/file_exchange_cgi.py
@@ -6,11 +6,11 @@ import logging
 import os
 
 
-logging.basicConfig(
-	filename="/usr/share/nginx/media/out.log",
-	format='%(asctime)s - %(levelname)s: %(message)s',
-	level=logging.DEBUG
-)
+# logging.basicConfig(
+# 	filename="/usr/share/nginx/media/out.log",
+# 	format='%(asctime)s - %(levelname)s: %(message)s',
+# 	level=logging.DEBUG
+# )
 
 class CgiComponents:
 	def __init__(self):
@@ -33,17 +33,49 @@ class Response:
 		self.status_code = 200
 		self.body = ""
 
-	def get_full_response(self):
+	def get_full_response(self) -> str:
 		code_response = {
-			200: self._successful_response
+			200: self._successful_response,
+			201: self._created_response,
+			400: self._bad_request_response,
+			405: self._method_not_allowed_response,
+			500: self._server_error_response
 		}
 		return code_response[self.status_code]()
 	
 	def _successful_response(self) -> str:
 		body = '<html><body>\n<h1>Python CGI "File Exchange"</h1>' \
 			   f'{self.body}</body></html>'
-		response = "HTTP/1.1 200 OK\nContent-Type: text/html\n" \
-				   f"Content-Length: {len(body)}\n\n{body}"
+		response = "Status: 200 OK\r\nContent-Type: text/html\r\n" \
+				   f"Content-Length: {len(body)}\r\n\r\n{body}"
+		return response
+	
+	def _created_response(self) -> str:
+		body = '<html><body>\n<h1>Python CGI "File Exchange"</h1>' \
+			   f'{self.body}</body></html>'
+		response = "Status: 201 Created\r\nContent-Type: text/html\r\n" \
+				   f"Content-Length: {len(body)}\r\n\r\n{body}"
+		return response
+	
+	def _bad_request_response(self) -> str:
+		body = '<html><body>\n<h1>Python CGI "File Exchange"</h1>' \
+			   f'<h2>400 Bad Request</h2>{self.body}</body></html>'
+		response = "Status: 400 Bad Request\r\nContent-Type: text/html\r\n" \
+				   f"Content-Length: {len(body)}\r\n\r\n{body}"
+		return response
+	
+	def _method_not_allowed_response(self) -> str:
+		body = '<html><body>\n<h1>Python CGI "File Exchange"</h1>' \
+			   '<h2>405 This method is not allowed</h2></body></html>'
+		response = "Status: 405 Method Not Allowed\r\nContent-Type: text/html\r\n" \
+				   f"Content-Length: {len(body)}\r\n\r\n{body}"
+		return response
+	
+	def _server_error_response(self) -> str:
+		body = '<html><body>\n<h1>Python CGI "File Exchange"</h1>' \
+			   f'<h2>500 Internal Server Error</h2>{self.body}</body></html>'
+		response = "Status: 500 Internal Server Error\r\nContent-Type: text/html\r\n" \
+				   f"Content-Length: {len(body)}\r\n\r\n{body}"
 		return response
 
 
@@ -57,61 +89,78 @@ class Cgi:
 	def handle_request(self):
 		method = os.environ.get("REQUEST_METHOD", "")
 		response = Response()
-		files = os.listdir(self.uploads_dir) # Should run in try-except
 		if method == "GET":
-			if not files:
-				response.body += "<p>No files here yet</p>"
-			else:
-				response.body += "<h2>File list:</h2>\n<ul>"
-				for file in files:
-					response.body += self.components.file_with_delete_button(file)
-				response.body += "</ul>"
-			response.body += self.components.upload_form
-			with open(self.delete_js_filepath, "r") as js_file:
-				js_script = js_file.read()
-			response.body += f"<script>\n{js_script}</script>"
+			self._get_main_page(response)
 		elif method == "POST":
-			logging.debug("Got post request")
-			if not self.storage:
-				logging.debug("Empty storage")
-			for key in self.storage:
-				logging.debug(f"Key: {key},\ndata:\n{self.storage[key]}")
-			file_object = self.storage["file"]
-			response.body += self._create_file(file_object)
+			self._create_file(response)
 		elif method == "DELETE":
-			filename = self.storage.getvalue("filename")
-			response.body += self._delete_file(filename)
+			self._delete_file(response)
 		else:
-			response.body += f"<h2>Unknown method: {method}</h2>"
+			response.status_code = 405
 		print(response.get_full_response())
 
-	def _delete_file(self, filename: str) -> str:
+	def _get_main_page(self, response: Response):
+		try:
+			files = os.listdir(self.uploads_dir)
+		except Exception as e:
+			response.status_code = 500
+			response.body = f"<p>Error receiving a list of files: {e}</p>"
+			return
+		if not files:
+			response.body = "<p>No files here yet</p>"
+		else:
+			response.body = "<h2>File list:</h2>\n<ul>"
+			for file in files:
+				response.body += self.components.file_with_delete_button(file)
+			response.body += "</ul>"
+		response.body += self.components.upload_form
+		with open(self.delete_js_filepath, "r") as js_file:
+			js_script = js_file.read()
+		response.body += f"<script>\n{js_script}</script>"
+
+	def _delete_file(self, response: Response):
+		filename = self.storage.getvalue("filename")
+		if filename is None:
+			response.status_code = 400
+			response.body = "<p>Field 'filename' is required<p>"
+			return
 		filepath = f"{self.uploads_dir}/{filename}"
 		if not os.path.isfile(filepath):
-			return f"<p>File '{filename}' does't exist</p>"
+			response.status_code = 400
+			response.body = f"<p>File '{filename}' does't exist</p>"
+			return
 		try:
 			os.remove(filepath)
-			return f"<p>File '{filename}' was deleted successfully</p>"
+			response.body = f"<p>File '{filename}' was deleted successfully</p>"
 		except Exception as e:
-			return f"<p>Error deleting file '{filename}': {e}</p>"
+			response.status_code = 500
+			response.body = f"<p>Error deleting a file '{filename}': {e}</p>"
 
-	def _create_file(self, file_object: cgi.FieldStorage) -> str:
+	def _create_file(self, response: Response):
+		if "file" not in self.storage:
+			response.status_code = 400
+			response.body = "<p>Field 'file' is required<p>"
+			return
+		file_object = self.storage["file"]
 		filename = os.path.basename(file_object.filename)
 		filepath = f"{self.uploads_dir}/{filename}"
 		if os.path.isfile(filepath):
-			return f"<p>File '{filename}' is already exist</p>"
+			response.status_code = 400
+			response.body = f"<p>File '{filename}' already exists</p>"
+			return
 		try:
 			with open(filepath, "wb") as file:
 				file.write(file_object.file.read())
-			return f"<p>File {filename} was uploaded</p>"
+			response.status_code = 201
+			response.body = f"<p>File {filename} was uploaded</p>"
 		except Exception as e:
-			return f"<p>Error creating file '{filename}': {e}</p>"
+			response.status_code = 500
+			response.body = f"<p>Error creating file '{filename}': {e}</p>"
 
 
 def run_cgi():
 	cgitb.enable()
 	cgi = Cgi()
-	logging.debug("Run CGI")
 	cgi.handle_request()
